requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.2
cryptography==45.0.5
cyclopts==3.22.2
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
email_validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.10.5
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp==1.11.0
mdurl==0.1.2
openapi-pydantic==0.5.1
packaging==25.0
pluggy==1.6.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyperclip==1.9.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rich-rst==1.3.1
rpds-py==0.26.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.4.1
starlette==0.47.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uv==0.7.21
uvicorn==0.35.0
